refactor: clean up DataBrowser code per review feedback
- Remove leftover SubscriptionManager wrapper class, use
  GranularSubscriptionManager directly
- Rename DataBrowserStore to ValueEmittingStore for clarity
- Remove unused clearContext() and clear() methods
- Remove implementation note from TimestampCell comment
- Simplify visibleItems loop to use map() instead of for-loop
